Move search use-cases to search module
