+++ appBackend/search-data/src/jsMain/kotlin/com/gchristov/thecodinglove/searchdata/api/ApiSearchResult.kt
@@ -1,6 +1,5 @@
 package com.gchristov.thecodinglove.searchdata.api
 
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -9,11 +8,4 @@ data class ApiSearchResult(
     val query: String,
     val post: ApiPost,
     val totalPosts: Int
-)
-
-fun SearchWithSessionUseCase.Result.toSearchResult() = ApiSearchResult(
-    searchSessionId = searchSessionId,
-    query = query,
-    post = post.toPost(),
-    totalPosts = totalPosts
 )
