+++ appBackend/search-testfixtures/build.gradle.kts
@@ -7,7 +7,7 @@ kotlin {
         val jsMain by getting {
             dependencies {
                 implementation(projects.commonServiceData)
-                implementation(projects.searchData)
+                implementation(projects.search)
             }
         }
     }
