+++ appBackend/search/src/jsMain/kotlin/com/gchristov/thecodinglove/search/SearchModule.kt
@@ -4,10 +4,13 @@ import com.gchristov.thecodinglove.commonservicedata.api.ApiServiceRegister
 import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSubSender
 import com.gchristov.thecodinglove.commonservicedata.pubsub.PubSubServiceRegister
 import com.gchristov.thecodinglove.kmpcommondi.DiModule
-import com.gchristov.thecodinglove.searchdata.usecase.PreloadSearchResultUseCase
-import com.gchristov.thecodinglove.searchdata.usecase.SearchWithSessionUseCase
+import com.gchristov.thecodinglove.search.usecase.*
+import com.gchristov.thecodinglove.searchdata.SearchRepository
+import com.gchristov.thecodinglove.searchdata.model.SearchConfig
+import kotlinx.coroutines.Dispatchers
 import kotlinx.serialization.json.Json
 import org.kodein.di.DI
+import org.kodein.di.bindProvider
 import org.kodein.di.bindSingleton
 import org.kodein.di.instance
 
@@ -16,6 +19,24 @@ object SearchModule : DiModule() {
 
     override fun bindDependencies(builder: DI.Builder) {
         builder.apply {
+            bindProvider {
+                provideSearchWithHistoryUseCase(
+                    searchRepository = instance(),
+                    searchConfig = instance()
+                )
+            }
+            bindProvider {
+                provideSearchWithSessionUseCase(
+                    searchRepository = instance(),
+                    searchWithHistoryUseCase = instance(),
+                )
+            }
+            bindProvider {
+                providePreloadSearchResultUseCase(
+                    searchRepository = instance(),
+                    searchWithHistoryUseCase = instance(),
+                )
+            }
             bindSingleton {
                 provideSearchApiService(
                     apiServiceRegister = instance(),
@@ -33,6 +54,33 @@ object SearchModule : DiModule() {
             }
         }
     }
+
+    private fun provideSearchWithHistoryUseCase(
+        searchRepository: SearchRepository,
+        searchConfig: SearchConfig
+    ): SearchWithHistoryUseCase = RealSearchWithHistoryUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchConfig = searchConfig
+    )
+
+    private fun provideSearchWithSessionUseCase(
+        searchRepository: SearchRepository,
+        searchWithHistoryUseCase: SearchWithHistoryUseCase
+    ): SearchWithSessionUseCase = RealSearchWithSessionUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchWithHistoryUseCase = searchWithHistoryUseCase
+    )
+
+    private fun providePreloadSearchResultUseCase(
+        searchRepository: SearchRepository,
+        searchWithHistoryUseCase: SearchWithHistoryUseCase
+    ): PreloadSearchResultUseCase = RealPreloadSearchResultUseCase(
+        dispatcher = Dispatchers.Default,
+        searchRepository = searchRepository,
+        searchWithHistoryUseCase = searchWithHistoryUseCase
+    )
 
     private fun provideSearchApiService(
         apiServiceRegister: ApiServiceRegister,
